reporter: fix bug around services disappearing after a crash
Scenario:

Nerve terminates prematurely (maybe because of a disconnection for a ZK
server).

It gets restarted by runit.

It then does its health checks, finds that the service is up.

It tries to create the node, fails because the node already exists,
so it sets its value.

The session for the previous instance of nerve expires, and the node
being ephemeral and attached to that session, disappears.

Nerve doesn't do anything about that unless the service goes down then
back up, at which point the node gets recreated.