log service watcher errors when not in debug mode
Currently, unless the log level is set to debug, nerve is silent as to _why_ a
service is marked unhealthy.  We should not have to spam the log files with
debug messages in order to find out why a service was marked down.

Also, move ignore_errors into the base watcher so that it can log the name of
the failing watcher, and rename it to catch_errors since ignoring isn't quite
what we're doing any more.

Also: log the response body when we get an error or loglevel is debug.