Add content check to HttpServiceCheck.
Introduces a new HttpServiceCheck configuration parameter called
'expect' which is string value to expect in the server response.

In addition to checking for a status code in the [200,300) range, it
also verifies a given string is contained in the response. This is similar
to the HAProxy 'http-check expect string' configuration.