Use pooling instead of shuffle for connection distribution
This makes it so that we better pool zookeeper connections for 5 node
zookeeper clusters (instead of 120 connections we only hold 1). I
verified with jnb that the zk client already randomly selects the host
to connect to so nerve does not need to do that now that we are re-using
connections.

This also fixes a bug with how I closed connections in the last pooling
implementation, hence the extra book-keeping.

No more randomization since we are pooling so well