Instead of fast failing, nerve should auto-remediate
Zookeeper is a CP store, not an AP store, so it flakes sometimes. In
production this means that we will often see networking flakes in one
datacenter impact another's registrations, and that's not ok. This patch
makes it so that instead of failing fast, nerve continuously tries to
revive dead watchers.