Add no-op service check
For when you just want to will something in to health.

This check can be used when the presence of a running nerve process on a
node is a reasonable enough proxy for likely health of the service the
node is meant to be running.  In our case, we're using nerve zk
registration to expose some services over DNS, and we have a service
that needs to have exposed DNS names _before_ the service can properly
start.  It is a special case where it's better to maintain the
registration than to accurately account for health.